rm_stm/txns: track term_id in begin_tx.
Now that transactions can span multiple terms, etag
check during prepare phase do not hold true anymore.

Rather we should compare it with the term_id in which
the transaction begin (begin_tx call).

Ideally we do not need this check anymore but we just
retain it for historical purposes. We can simplify it
in future patches.