serde: add support for std::chrono::duration
Generalizes support for duration and removes special
casing for std::chrono::milliseconds.

The patch serializes by casting to the equivalent
chrono::nanosecond, so this introduces an upper limit
of nanoseconds::max() (~292 years with int64_t rep)
on the duration. If the cast oveflows/underflows
the value is clamped to max/min nanoseconds respectively.