compaction: Fix concatention with empty segment.
Transaction fuzzer identified this bug in compaction code. In the
following sequence generated by the fuzzer.

-> roll
-> aborted ranges
-> roll

The segment in the middle will be empty after tx_reducer filters out
the aborted ranges and the offsets of the concatenated segment are
incorrect. This does not seem to be an issue with regular rolling of
log because we have explicit checks to filter out empty segment before
we roll to the new segment.