txs/ducky: remove test_delete_topic_from_prepared_tx
With a reliable way of detecting deleted topics, the
logic in the test does not hold true anymore.
The admin API to remove a deleted partition from a tx
seems to have been done to workaroud the lack of deletion
support which we now have. We can still retain the API
as a fail safe option but removing the test for now.