tx/compaction: additional debug logging for missing aborted txs
Adds additional error logging when we come across an abort batch without
an equivalent aborted tx range. Ideally this should never happen and
we need to error out but that needs a bigger refactor of code, will do
it as a separate patch. For now the error logging tells us the approx
offset at which this happened.