observability: Add a memory tracking allocator
Implements an std::allocator that tracks allocated and deallocated
memory bytes. Each allocator instance is tagged with a mem_tracker
instance.

mem_trackers can form a hierarchy, used to compute the memory breakdown.

ex: We can have a root tracker for the redpanda process and child
trackers for each of the subsystems.

Utilities are implemented to compute the consumption for the entire
tree rooted at a tracker and to pretty print the tree in ascii.