c/backend: unconditionally compute bootstrap revisions
This is a revert of d3ce3cefe.

d3ce3cefe added coordination between shards for x core movements during
bootstrap. The approach implicitly assumed that cancled x core
operations are not executed during bootstrap.

This assumption does not hold after 8b500eab322b because each shard
executes every operation irrespective of whether it was canceled or
not to factor in cancelation reverts.

This commit temporarily reverts the changes in d3ce3cefe and the next
set of commits fix the underlying issue.