c/backend: Differentiate current revision vs update revisions
In case of x core updates, always using the current revision of the ntp
can result in incorrect updates to the shard table. This happens because
each shard bootstraps the controller log in parallel and depending on
the scheduling points, it can generate multiple interleaved update
possibilities causing incorrect shard table entries.

This patch differentiates between the current revision vs the revision
of the update (= offset of the delta in the log for that update)  that
is currently executing. We use the latter to update the shard table as
that guarantees that newer revisions take precendence over older ones.