rm_stm: cleanup snapshot related code.
No logical changes in this patch.

In the spirit of making state machine leaner, moves a ton of
definitions from rm_stm header into its its own header. Most of this
code will be deprecated once transactional state is ported to
producer_state _and_ we switch to serde for serialization. Parking
the structs in this header until then.