rm_stm: Add producers map
This map tracks all the producers producing to this raft group.
Producers can be both idempotent/transactional. (non idempotent
producers have no PID).

As noted in earlier commits, eventually this is the _only_ map that
should exist in the stm. But becauase we are porting only idempotency
code to the producer_state, this map needs to coexist with other
transactional state until the porting is done.