rm_stm/idempotency: add unknown_producer_id
If a producer_id is not known (which is supposed to be known) this
commit makes the broker throw uknown_producer_id error instead of OOOSN
which forces the client to reset epoch and try again.

Not all clients (eg: librdkafka, kgo) reset epoch on OOOSN but without
inflight requests. Only Java client seems to do that. All clients seem
to consistently handle unknown_producer_id by reseting the epoch and
trying again.

Java client requires log_start_offset to be set in the produce_response
to function, which is done by this patch.