datalake/translation: introduce partition_translator
A partition translator is responsible for translating data from a given
partition periodically. A single instance of translator should be active for
a given partition and is attached to the leader replica.

The translator periodically wakes up, checks if there is new data that is
pending translation and creates a reader for the data range and hands it off
to low level iceberg translator that converts the data.