cluster_link: pass a unique_ptr<connection> for link creation
Future commits will need access to the cluster connection object (to
create a direct consumer). To enable this, this commit  uses a unique_ptr,
which allows referencing the object before passing it to the link.
This requirement only exists in the source implementation
(not in the API), so the connection must be set up during creation
while wiring things together. An alternative would be to pass the link
pointer into the data source factory, but that introduces an awkward
dependency cycle, this approach seems less ugly in comparison.