cl/link: implement status reconciliation
This reconciler is responsible to reconcile the status of topics from
transitional states to target states. Transitional states are
temporary states that are to be switched to a desired state. For example
failing_over -> failed_over
promoting -> promoted

In future we can also transitions from failed_over-> active with some
transitional state called activating.